Removed int enum type so that all sizes are know at generation-time in C++

+++ generation/dnp3/src/main/scala/com/automatak/render/dnp3/enums/ChannelState.scala
@@ -7,7 +7,7 @@ object ChannelState {
   private val comments = List("Enumeration for possible states of a channel")
   private val defaultValue = EnumValue("SHUTDOWN", 4, "stopped and will never do anything again")
 
-  def apply(): EnumModel = EnumModel("ChannelState", comments, EnumModel.Integer, codes, Some(defaultValue), Base10)
+  def apply(): EnumModel = EnumModel("ChannelState", comments, EnumModel.UInt8, codes, Some(defaultValue), Base10)
 
   private val codes = List(
     EnumValue("CLOSED", 0, "offline and idle"),
